fix(ontology): share entity-label classification across all ingest paths
The handler ingest path (process_experience_into_graph) classified MISC NER
entities via classify_misc_entity and tags via classify_tag_label, but the
core MemorySystem::upsert create-path used resolve_entity_label, which mapped
everything that was not PER/ORG/LOC to Concept. The same name therefore got a
rich label when created via remember and a flat Concept when created via an
integration sync or upsert — making it invisible or coarser to type-aware
spreading activation and matches_with_hierarchy() in retrieval.

- Move classify_tag_label and classify_misc_entity into graph_memory (next to
  EntityLabel) as pub fns; both ingest layers now call the one implementation.
- resolve_entity_label routes MISC and no-NER-hit names through
  classify_misc_entity instead of blanket Concept, so PER/ORG/LOC map directly
  and everything else resolves to the same heuristic label as the handler path.

Tests: classifier behavior locked at the definition site (graph_memory) and
the upsert path proven to no longer collapse to Concept (resolve_entity_label
tests). Existing handler-side classifier tests still pass against the moved
functions.
