+++ src/handlers/state.rs
@@ -414,167 +414,6 @@ fn issue_regex() -> &'static regex::Regex {
 /// `Technology`, this uses suffix/keyword matching to assign the correct
 /// ontological type — enabling type-aware spreading activation and
 /// `matches_with_hierarchy()` in retrieval.
-fn classify_tag_label(tag: &str) -> EntityLabel {
-    let lower = tag.to_lowercase();
-
-    // Deployment / environment indicators
-    if matches!(
-        lower.as_str(),
-        "production"
-            | "staging"
-            | "dev"
-            | "development"
-            | "ci"
-            | "cd"
-            | "kubernetes"
-            | "k8s"
-            | "docker"
-            | "container"
-            | "aws"
-            | "gcp"
-            | "azure"
-    ) {
-        return EntityLabel::Environment;
-    }
-
-    // Pipeline / workflow indicators
-    if lower.contains("pipeline")
-        || lower.contains("workflow")
-        || lower.contains("ci-cd")
-        || lower.contains("cicd")
-    {
-        return EntityLabel::Pipeline;
-    }
-
-    // Database / storage indicators
-    if matches!(
-        lower.as_str(),
-        "rocksdb"
-            | "postgres"
-            | "postgresql"
-            | "redis"
-            | "sqlite"
-            | "mongodb"
-            | "mysql"
-            | "dynamodb"
-            | "s3"
-            | "cassandra"
-            | "elasticsearch"
-    ) || lower.ends_with("db")
-        || lower.ends_with("-db")
-        || lower.ends_with("_db")
-    {
-        return EntityLabel::Database;
-    }
-
-    // Service / API indicators (suffix-only to avoid "my-api-docs" false positives)
-    if lower.ends_with("-service")
-        || lower.ends_with("_service")
-        || lower.ends_with("-api")
-        || lower.ends_with("_api")
-        || lower.ends_with("-server")
-        || lower.ends_with("_server")
-        || lower.ends_with("-daemon")
-    {
-        return EntityLabel::Service;
-    }
-
-    // Documentation indicators (check before module — README.md is a doc, not a module)
-    if lower.ends_with(".md")
-        || lower.contains("readme")
-        || lower.contains("runbook")
-        || lower.ends_with("-rfc")
-        || lower.ends_with("-spec")
-    {
-        return EntityLabel::Document;
-    }
-
-    // Configuration indicators
-    if lower.ends_with(".toml")
-        || lower.ends_with(".yaml")
-        || lower.ends_with(".yml")
-        || lower.ends_with(".env")
-        || lower.ends_with(".json")
-        || lower.contains("config")
-    {
-        return EntityLabel::Configuration;
-    }
-
-    // Module / library indicators
-    if lower.ends_with(".rs")
-        || lower.ends_with(".ts")
-        || lower.ends_with(".js")
-        || lower.ends_with(".py")
-        || lower.ends_with("-lib")
-        || lower.ends_with("_lib")
-    {
-        return EntityLabel::Module;
-    }
-
-    // Default: Technology (reasonable fallback for unknown tags)
-    EntityLabel::Technology
-}
-
-/// Classify a MISC NER entity into a more specific EntityLabel.
-///
-/// The NER model only outputs PER/ORG/LOC/MISC. For MISC entities that pass
-/// quality gates, we use heuristic rules to assign a richer ontological type.
-/// This activates type-aware spreading activation and matches_with_hierarchy()
-/// in retrieval — `Concept` entities participate in the type hierarchy while
-/// `Other("MISC")` is invisible to it.
-fn classify_misc_entity(name: &str) -> EntityLabel {
-    let lower = name.to_lowercase();
-
-    // Event-like terms (checked first — "conference" ends with "ence" concept suffix)
-    const EVENT_WORDS: &[&str] = &[
-        "conference",
-        "summit",
-        "meetup",
-        "hackathon",
-        "sprint",
-        "launch",
-        "release",
-        "incident",
-        "outage",
-        "postmortem",
-    ];
-    if EVENT_WORDS.iter().any(|w| lower.contains(w)) {
-        return EntityLabel::Event;
-    }
-
-    // Skill / role indicators (before concept — "engineer" ends with suffix-like patterns)
-    const SKILL_WORDS: &[&str] = &[
-        "engineer",
-        "architect",
-        "developer",
-        "designer",
-        "analyst",
-        "programming",
-        "scripting",
-    ];
-    if SKILL_WORDS.iter().any(|w| lower.contains(w)) {
-        return EntityLabel::Skill;
-    }
-
-    // Concept suffixes — abstract ideas, methodologies, qualities
-    const CONCEPT_SUFFIXES: &[&str] = &[
-        "ism", "ology", "ity", "ness", "ment", "ance", "ence", "tion", "sion",
-    ];
-    if CONCEPT_SUFFIXES.iter().any(|s| lower.ends_with(s)) {
-        return EntityLabel::Concept;
-    }
-
-    // PascalCase proper nouns without spaces/hyphens → likely a Product
-    // (e.g., "JavaScript", "PostgreSQL", "FastAPI", "ChatGPT")
-    let has_internal_upper = name.chars().skip(1).any(|c| c.is_uppercase());
-    if has_internal_upper && !name.contains(' ') && !name.contains('-') && name.len() > 2 {
-        return EntityLabel::Product;
-    }
-
-    // Default: Concept (participates in type hierarchy via Role→Concept parent)
-    EntityLabel::Concept
-}
-
 use crate::ab_testing;
 use crate::backup;
 use crate::config::ServerConfig;
@@ -583,8 +422,8 @@ use crate::embeddings::{
     KeywordExtractor, NerConfig, NeuralNer,
 };
 use crate::graph_memory::{
-    EdgeTier, EntityLabel, EntityNode, EpisodeSource, EpisodicNode, GraphMemory, GraphStats,
-    LtpStatus, RelationshipEdge,
+    classify_misc_entity, classify_tag_label, EdgeTier, EntityLabel, EntityNode, EpisodeSource,
+    EpisodicNode, GraphMemory, GraphStats, LtpStatus, RelationshipEdge,
 };
 use crate::memory::{
     Experience, FeedbackStore, FileMemoryStore, MemoryConfig, MemoryId, MemoryStats, MemorySystem,
